backend/requirements.txt
fastapi
uvicorn[standard]
feedparser
newspaper3k 